Fix reference before declaration
